# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: |
    Internal static website hosting. Every site is served at the root of its
    own host, `<sitename>.<owner>.<base>/` (for example
    `todo.alice.example.com/`). The owner host `<owner>.<base>/` serves the
    person's or team's index page. Until the owner's certificate is ready it
    also serves their sites at `<owner>.<base>/<sitename>/`; after that,
    those addresses, and `/api/sites/<sitename>/...` on the owner host,
    redirect to each site's own host with path and query kept (`301` for GET
    and HEAD, `308` otherwise). v1.2 addresses of the form
    `<owner>--<sitename>.<base>` only redirect. Always use the `url` from the
    latest response rather than composing an address. Uploads may
    be gzip-compressed tar or ZIP archives; nested regular-file downloads such
    as ZIP and DMG artifacts are supported.

    A site is owned by a person or a team. Team members change a team's
    sites through owner-qualified collaboration routes with optimistic
    `ETag` / `If-Match` protection.

    Every site has an access level: `only_me` (the owner, or the team's
    members; the default for a new site), `specific` (plus named viewers),
    `company` (anyone signed in with the link),
    `listed` (company, plus showcase and search), or `network` (anyone who
    can reach the server, no sign-in). `network` is requested with a reason
...
    network site's pages, assets and state, and change nothing.

...

...

...

...
servers:
...
paths:
...
  /skills/sha256/{digest}/skills.zip:
    get:
      summary: Download a bundle by content digest
      description: Serves the skill bundle only when {digest} is the SHA-256 of the bundle this deployment produces; any other digest returns 404. Because the URL names the bytes, the response is safely cacheable forever. Read immutable_bundle_url from /skills/version rather than constructing this path.
      tags: [Skills]
      security: []
      parameters:
        - name: digest
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{64}$'
      responses:
        '200':
          description: The bundle matching this digest
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [public, max-age=31536000, immutable]
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          description: No deployed bundle has this digest. Not cached, since a later deployment may serve it.
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        An API key, for CI and other automation; agents normally connect
        through `/mcp` with the company sign-in instead. Each key has a scope
        chosen when it is created and an expiry (at most 365 days).
        `publish` (the default) deploys, updates and rolls back sites, reads
        versions, saved data and assets, and calls `/api/me` and `/mcp`.
        `full` adds everything else the person can do (deleting sites, access
        levels, viewers, teams, the audit and access logs, search), except
        administration. `offboard` (admins only) calls
        `POST /api/admin/users/disable` and nothing else. No key reaches any
        other `/api/admin/*` route. OAuth connector tokens work only on
        `/mcp`.
...